refactor: extract left menu subcomponents
